Added logging to pipeline

+++ pipeline/Run_pipeline.py
@@ -6,12 +6,14 @@ import os
 from dotenv import load_dotenv
 import requests
 import gzip
-from tqdm.notebook import trange
+import logging
+from pipeline_logger import setup_logger
+
+logger = logging.getLogger("imdb_pipeline.run_pipeline")
 
 # Config
 PROJECT_ROOT = Path(__file__).resolve().parents[1]
 load_dotenv(PROJECT_ROOT / ".env")
-
 COMPOSE_FILE = PROJECT_ROOT / "docker" / "docker-compose.yml"
 SQL_DIR = PROJECT_ROOT / "mysql"
 DATA_DIR = PROJECT_ROOT / "data" / "tests"
@@ -21,11 +23,12 @@ DOCKER = "docker"
 MYSQL_PASSWORD = os.getenv("MYSQL_ROOT_PASSWORD")
 
 if not MYSQL_PASSWORD:
-    raise RuntimeError(
+    logger.error(
         "MYSQL_ROOT_PASSWORD not set.\n"
         "Create a .env file in project root.\n"
         "See .env.example for required variables."
     )
+    raise RuntimeError("Missing required environment variable: MYSQL_ROOT_PASSWORD")
 
 env = os.environ.copy()
 env["MYSQL_PWD"] = MYSQL_PASSWORD
@@ -35,7 +38,7 @@ env["MYSQL_PWD"] = MYSQL_PASSWORD
 
 def run_command(cmd, check=True):
     """Run command"""
-    print(f"\nRunning: {' '.join(map(str, cmd))}")
+    logger.info(f"\nRunning: {' '.join(map(str, cmd))}")
 
     result = subprocess.run(
         cmd,
@@ -44,9 +47,10 @@ def run_command(cmd, check=True):
     )
 
     if result.returncode != 0:
-        print("STDOUT:\n", result.stdout)
-        print("STDERR:\n", result.stderr)
+        logger.warning("STDOUT:\n%s", result.stdout)
+        logger.warning("STDERR:\n%s", result.stderr)
         if check:
+            logger.error(f"Command failed with exit code {result.returncode}")
             raise RuntimeError(
                 f"Command failed with exit code {result.returncode}"
             )
@@ -66,24 +70,28 @@ def download_files():
     for url in urls:
         filename = url.split('/')[-1]
         target_path = PROJECT_ROOT / "data" / "raw" / filename
-        print(target_path)
-        response = requests.get(url, stream=True)
+        logger.info(target_path)
 
-        if response.status_code == 200:
-            with open(target_path, 'wb') as f:
-                f.write(response.raw.read())
-        target_path = Path(target_path)
+        if target_path.exists():
+            logger.info("File %s already exists. Skipping download.", target_path)
+        else:
+            response = requests.get(url, stream=True)
+
+            if response.status_code == 200:
+                with open(target_path, 'wb') as f:
+                    f.write(response.raw.read())
 
         if target_path.suffix == ".gz":
             tsv_path = target_path.with_suffix("")
 
         with gzip.open(target_path, 'rb') as gz_file:
-            with open(tsv_path, 'wb') as f:
-                f.write(gz_file.read())
-
-        for i in trange(1, desc='Statut'):
-            print('Fichier téléchargé :', tsv_path)
-
+            if tsv_path.exists():
+                logger.info("File %s already exists. Skipping extraction.", tsv_path)
+            else:
+                logger.info("Extracting file to %s", tsv_path)
+                with open(tsv_path, 'wb') as f:
+                    f.write(gz_file.read())
+                logger.info("Extraction completed for %s", tsv_path)
 
 def copy_tsv_to_container():
     file_names = [
@@ -94,7 +102,7 @@ def copy_tsv_to_container():
         "title.ratings.tsv"]
 
     for file in file_names:
-        print(f"Copying {file} into docker container")
+        logger.info(f"Copying {file} into docker container")
         result = subprocess.run(["docker",
                                  "cp",
                                  f"{PROJECT_ROOT}/data/raw/{file}",
@@ -102,56 +110,11 @@ def copy_tsv_to_container():
                                 capture_output=True,
                                 text=True,
                                 check=False)
-        print(result.stderr)
-
-
-def wait_for_mysql(container, timeout=60):
-    """Wait until MySQL inside container is ready."""
-    print("Waiting for MySQL to be ready...")
-    start = time.time()
-    subprocess.run(["docker", "ps"])
-
-    while time.time() - start < timeout:
-        # result = subprocess.run(
-        #     [DOCKER, "exec", "-e",
-        #      f"MYSQL_PWD={MYSQL_PASSWORD}",
-        #      container,
-        #      "mysqladmin",
-        #      "-u", "root",
-        #      "ping"],
-        #     capture_output=True,
-        #     text=True
-        # )
-        result = subprocess.run(
-            [
-                DOCKER, "exec",
-                "-e", f"MYSQL_PWD={MYSQL_PASSWORD}",
-                CONTAINER,
-                "mysqladmin",
-                "-h", "127.0.0.1",
-                "-u", "root",
-                "ping"
-            ],
-            capture_output=True,
-            text=True
-        )
-
-        print("STDOUT:", result.stdout.strip())
-        print("STDERR:", result.stderr.strip())
-        print("RETURN:", result.returncode)
-
-        if "mysqld is alive" in result.stdout:
-            print("MySQL is ready.")
-            return
-
-        time.sleep(1)
-
-    raise TimeoutError("MySQL did not become ready in time.")
-
+        logger.debug(result.stderr)
 
 def execute_sql_file(sql_file):
     """Execute one SQL file inside container."""
-    print(f"\nExecuting SQL file: {sql_file.name}")
+    logger.info(f"Executing SQL file: {sql_file.name}")
 
     # Remove CSV file if it exists in container
     start = time.time()
@@ -164,7 +127,7 @@ def execute_sql_file(sql_file):
             text=True
         )
         if sock.returncode == 0 and "ready" in sock.stdout:
-            print("MySQL socket is ready.")
+            logger.info("MySQL socket is ready.")
             break
     with open(sql_file, "rb") as f:
         result = subprocess.run(
@@ -181,12 +144,12 @@ def execute_sql_file(sql_file):
     if result.returncode != 0:
         raise RuntimeError(f"SQL execution failed for {sql_file.name}")
 
-    print(f" Finished running {sql_file.name}")
+    logger.info(f" Finished running {sql_file.name}")
 
 
 def cleanup():
     """Stop and remove Docker container."""
-    print("Cleaning up: Stopping Docker container...")
+    logger.info("Cleaning up: Stopping Docker container...")
     run_command([DOCKER, "compose", "-f", str(COMPOSE_FILE), "down"])
 
 
@@ -226,15 +189,15 @@ def update_csv_file(file):
             names=columns)
 
     marvel_df.to_csv(file, index=False)
-    print(marvel_df.head())
-    print(marvel_df.shape)
-    print("Returned rows:", len(marvel_df))
+    logger.debug(marvel_df.head())
+    logger.debug(marvel_df.shape)
+    logger.info("Returned rows: %d", len(marvel_df))
 
 # Pipeline execution
 
 
 def main():
-
+    logger = setup_logger()
     # Start Docker container
     try:
         run_command([
@@ -242,9 +205,9 @@ def main():
             "-f", str(COMPOSE_FILE),
             "up", "-d"
         ])
-        print("Docker container started successfully.")
+        logger.info("Docker container started successfully.")
     except Exception as e:
-        print(f"Error starting Docker container: {e}")
+        logger.error(f"Error starting Docker container: {e}")
         return
 
     # Download files from IMDB
@@ -257,9 +220,6 @@ def main():
                       "imdb-load-data.sql", "imdb-add-constraints.sql",
                       "imdb-add-index.sql"]
 
-    #  Wait for MySQL to be ready
-    wait_for_mysql(CONTAINER)
-
     # Execute SQL files to create DB, tables, load data, add constraints
     # and indexes
     for sql in db_setup_files:
@@ -268,6 +228,7 @@ def main():
     #  Get SQL files
     sql_files = sorted(SQL_DIR.glob("*marvel*.sql"))
     if not sql_files:
+        logger.error("No SQL files found in %s", SQL_DIR)
         raise FileNotFoundError("No SQL files found.")
 
     # Remove any existing CSV files in container before running queries that
@@ -278,7 +239,7 @@ def main():
     for sql_file in sql_files:
         execute_sql_file(sql_file)
 
-    print("\n All SQL queries executed successfully.")
+    logger.info("All SQL queries executed successfully.")
 
     for csv in DATA_DIR.glob("*marvel*.csv"):
         update_csv_file(csv)
